Added end effector design of Carbonoid D1
The end effector design is derived from the work I did for the OpenBeam
KosselPro. The Carbonoid D1 end effector is designed for the E3D
Volcano hotend, however the E3D V6 will also work, but then the
filament fans must be adapted. The design supports 1.75mm and 3mm
versions of the hotend - only the hotend clip is different.